requirements.txt
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dateutil>=2.8.2